requirements.txt
coverage==7.10.7
fastapi==0.117.1
flake8==7.3.0
pytest==8.4.2
pytest-cov==7.0.0
uvicorn==0.37.0
aio_pika==9.5.7
pytest-asyncio==0.21.0
